Clean up data fetch, abstract serialization
